cleanup after code generation
