+++ client/src/app/pages/user-management/users/user-modal.tsx
@@ -14,22 +14,21 @@ import { useUserForm } from "./use-user-form";
 import { UserForm } from "./user-form";
 
 export interface UserModalProps {
-  isOpen: boolean;
   user?: User;
   onClose: () => void;
 }
 
-export const UserCreateModal: FC<Omit<UserModalProps, "user">> = ({
+// re-create the modal to reset all state
+export const UserCreateModal: FC<UserModalProps & { isOpen: boolean }> = ({
   isOpen,
   onClose,
-}) => <UserModal isOpen={isOpen} onClose={onClose} />;
+}) => isOpen && <UserModal onClose={onClose} />;
 
-export const UserEditModal: FC<Omit<UserModalProps, "isOpen">> = ({
-  user,
-  onClose,
-}) => <UserModal isOpen={!!user} user={user} onClose={onClose} />;
+// re-create the modal to reset all state
+export const UserEditModal: FC<UserModalProps> = ({ user, onClose }) =>
+  !!user && <UserModal user={user} onClose={onClose} />;
 
-const UserModal: FC<UserModalProps> = ({ isOpen, user, onClose }) => {
+const UserModal: FC<UserModalProps> = ({ user, onClose }) => {
   const { t } = useTranslation();
 
   const { form, onSubmit, isSubmitDisabled, isEdit, isSeeded } = useUserForm(
@@ -38,35 +37,33 @@ const UserModal: FC<UserModalProps> = ({ isOpen, user, onClose }) => {
   );
   return (
     <>
-      {isOpen && (
-        <Modal isOpen onClose={onClose} variant="medium">
-          <ModalHeader
-            title={user ? t("titles.editUser") : t("titles.createUser")}
+      <Modal isOpen onClose={onClose} variant="medium">
+        <ModalHeader
+          title={user ? t("titles.editUser") : t("titles.createUser")}
+        />
+        <ModalBody>
+          <UserForm
+            form={form}
+            isEdit={isEdit}
+            isSeeded={isSeeded}
+            onClose={onClose}
+            onSubmit={onSubmit}
           />
-          <ModalBody>
-            <UserForm
-              form={form}
-              isEdit={isEdit}
-              isSeeded={isSeeded}
-              onClose={onClose}
-              onSubmit={onSubmit}
-            />
-          </ModalBody>
-          <ModalFooter>
-            <Button
-              key="confirm"
-              variant="primary"
-              isDisabled={isSubmitDisabled}
-              onClick={onSubmit}
-            >
-              {user ? t("actions.save") : t("actions.create")}
-            </Button>
-            <Button key="cancel" variant="link" onClick={onClose}>
-              {t("actions.cancel")}
-            </Button>
-          </ModalFooter>
-        </Modal>
-      )}
+        </ModalBody>
+        <ModalFooter>
+          <Button
+            key="confirm"
+            variant="primary"
+            isDisabled={isSubmitDisabled}
+            onClick={onSubmit}
+          >
+            {user ? t("actions.save") : t("actions.create")}
+          </Button>
+          <Button key="cancel" variant="link" onClick={onClose}>
+            {t("actions.cancel")}
+          </Button>
+        </ModalFooter>
+      </Modal>
     </>
   );
 };
